top-songs-backend/pyproject.toml
[tool.poetry]
name = "top-songs-backend"
version = "0.1.0"
description = ""
authors = ["Esma BEN HABIB <esma.benhabib@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
qdrant-client = "^1.11.0"
fastapi = "^0.112.1"
pydantic = "^2.8.2"
sentence-transformers = "^3.0.1"
uvicorn = "^0.30.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
